fix(bugs): TC1121 TC854 TC1132 TC685 TC686 and more

+++ packages/sgcart/product-variants/src/Http/Controllers/ColorController.php
@@ -67,6 +67,10 @@ class ColorController extends Controller
 
     public function destroy(Color $color)
     {
+        if ($color->variants()->whereHas('product')->exists()) {
+            return redirect()->route('admin.colors.index')->with('error',
+                'Color swatch cannot be deleted because it is associated with a product variant.');
+        }
         $color->delete();
         return redirect()->route('admin.colors.index')->with('success', 'Color swatch deleted successfully.');
     }
